refactor: movendo loadingTimeout para pasta utils
